+++ PRN232_Assignment1/DTO/Request/ProductSearchRequestDto.cs
@@ -6,11 +6,11 @@ public class ProductSearchRequestDto
 {
     public string? SearchTerm { get; set; }
     
-    [Range(0, double.MaxValue, ErrorMessage = "Min price must be greater than or equal to 0")]
-    public decimal? MinPrice { get; set; }
+    [Range(0, float.MaxValue, ErrorMessage = "Min price must be greater than or equal to 0")]
+    public float? MinPrice { get; set; }
     
-    [Range(0, double.MaxValue, ErrorMessage = "Max price must be greater than or equal to 0")]
-    public decimal? MaxPrice { get; set; }
+    [Range(0, float.MaxValue, ErrorMessage = "Max price must be greater than or equal to 0")]
+    public float? MaxPrice { get; set; }
     
     [Range(1, int.MaxValue, ErrorMessage = "Page number must be greater than 0")]
     public int Page { get; set; } = 1;
